refactor: Move authOptions to lib/auth.ts for better organization
